cgroups: remove unused files
they are a dup of the _linux.go version.  Move the linux-only
functions to the _linux.go file.

[NO NEW TESTS NEEDED]
